requirements.txt
# requirements.txt
torch>=2.0.0
chess>=1.999
berserk>=0.13.0
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
